+++ src/file-system/current-directory-provider.class.ts
