+++ src/index.ts
@@ -1,5 +1,5 @@
 import Path from 'path';
-import { CurrentDirectoryProvider } from './file-system/current-directory-provider';
+import { CurrentDirectoryProvider } from './file-system/current-directory-provider.class';
 import { DirectoryInfo } from './file-system/directory-info.class';
 import { MiniBackup } from './mini-backup';
 import { Config } from './models/config.type';
